requirements.txt
pysc2
numpy>=1.14
gym>=0.12.0
pylint>=2.3.1
pytest>=4.4.0
absl-py==0.7.0